rename: RequestTokenExtractor -> TokenExtractor
